[2/4] DXCDT-266: Bring branding emails command under email templates (#540)
